ASN1 Decode and encode Ran Functions Accepted and Ran Functions rejected in E2SetupResponse (#85)
* [SDRAN-160] Handle processing of RanFunctionsIDs

* Added in RANfunctionIDCause encoding

* Added in RanFunctionIDCause decoding

* addressed warning in Manager Test

+++ pkg/southbound/e2ap/pdubuilder/e2-setup-request.go
@@ -56,7 +56,7 @@ func CreateE2SetupRequestPdu(plmnID types.PlmnID, ranFunctionIds types.RanFuncti
 				Criticality: int32(e2ap_commondatatypes.Criticality_CRITICALITY_IGNORE),
 				Value: &e2appducontents.RanfunctionItem{
 					RanFunctionId: &e2apies.RanfunctionId{
-						Value: id,
+						Value: int32(id),
 					},
 					RanFunctionDefinition: &e2ap_commondatatypes.RanfunctionDefinition{
 						Value: []byte(ranFunctionID.Description),
